Time: 1 ms (99.98%), Space: 42.7 MB (46.17%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -16,6 +16,6 @@ class Solution {
             }
      
 
-      return (int)rev;
+      return rev;
     }
 }
